+++ config/websocket.go
@@ -16,39 +16,42 @@ const (
 	DefaultPingPeriod = (DefaultPongTimeout * 9) / 10
 	// DefaultMaxMessageSize 1024
 	DefaultMaxMessageSize = 1024
-	// DefaultAutoStart true
-	DefaultAutoStart = true
 )
 
 //
 
-// Ws the config contains options for 'ws' package
-type Ws struct {
+// Websocket the config contains options for 'websocket' package
+type Websocket struct {
 	// WriteTimeout time allowed to write a message to the connection.
+	// Default value is 10 * time.Second
 	WriteTimeout time.Duration
 	// PongTimeout allowed to read the next pong message from the connection
+	// Default value is 60 * time.Second
 	PongTimeout time.Duration
 	// PingPeriod send ping messages to the connection with this period. Must be less than PongTimeout
+	// Default value is (PongTimeout * 9) / 10
 	PingPeriod time.Duration
 	// MaxMessageSize max message size allowed from connection
+	// Default value is 1024
 	MaxMessageSize int
-	// AutoStart starts the websocket server automatically
-	AutoStart bool
+	// Endpoint is the path which the websocket server will listen for clients/connections
+	// Default value is empty string, if you don't set it the Websocket server is disabled.
+	Endpoint string
 }
 
-// DefaultWs returns the default config for iris-ws
-func DefaultWs() Ws {
-	return Ws{
+// DefaultWebsocket returns the default config for iris-ws websocket package
+func DefaultWebsocket() Websocket {
+	return Websocket{
 		WriteTimeout:   DefaultWriteTimeout,
 		PongTimeout:    DefaultPongTimeout,
 		PingPeriod:     DefaultPingPeriod,
 		MaxMessageSize: DefaultMaxMessageSize,
-		AutoStart:      DefaultAutoStart,
+		Endpoint:       "",
 	}
 }
 
 // Merge merges the default with the given config and returns the result
-func (c Ws) Merge(cfg []Ws) (config Ws) {
+func (c Websocket) Merge(cfg []Websocket) (config Websocket) {
 
 	if len(cfg) > 0 {
 		config = cfg[0]
@@ -62,7 +65,7 @@ func (c Ws) Merge(cfg []Ws) (config Ws) {
 }
 
 // MergeSingle merges the default with the given config and returns the result
-func (c Ws) MergeSingle(cfg Ws) (config Ws) {
+func (c Websocket) MergeSingle(cfg Websocket) (config Websocket) {
 
 	config = cfg
 	mergo.Merge(&config, c)